Break up run_sampling.rb into two scripts, the former which is called from the commandline, and run_sampling_lib.rb which allows the RunSampling class to be called from other ruby scripts.
